+++ src/hooks/precompact.ts
@@ -11,6 +11,7 @@
 import { HookInput, HookOutput } from "../core/types.js";
 import { ContextArchiver } from "../core/archiver.js";
 import { Logger } from "../utils/logger.js";
+import { getStoragePath } from "../utils/path-resolver.js";
 import * as path from "path";
 import * as fs from "fs";